Произведена модуляция файлов
