+++ js/check-length-str.js
@@ -1,0 +1,7 @@
+const checkLengthOfStr = function(str, maxLength = 140) {
+  return Number(str.length) <= maxLength;
+};
+
+// window.console.log(checkLengthOfStr(abc));
+
+export {checkLengthOfStr};
