+++ js/utils.js
@@ -1,0 +1,15 @@
+const abc = 'asdsafdfdfdfgfdgh';
+const OBJECT_COUNT = 25;
+const OBJECT_COMMENT_COUNT = 4;
+const NAMES = ['Дима','Антон','Сергей','Артем','Юра'];
+const COMMENTS = [' Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+]
+const DESCRIPTION = ['Кот-суши', 'Закат', 'Озеро', 'Очень красивые глаза', 'Супер-кар'];
+
+
+export {abc, OBJECT_COUNT, OBJECT_COMMENT_COUNT, NAMES, COMMENTS, DESCRIPTION}
